pyproject.toml
[project]
name = "lambda-inspired-platform"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "rich (>=14.0.0,<15.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "kafka-python (>=2.2.15,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
